Move Interface into the same Namespace
