+++ 01_CODE_Frontend/Modules/HeaderModule.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using CODE_Frontend.ViewModel;
-using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.Entity;
-using CODE_GameLib.Interfaces.RoomObjects.Wearable;
+using CODE_GameLib.Objects;
+using CODE_GameLib.Objects.Entity;
+using CODE_GameLib.Objects.RoomObjects.Wearable;
 
 namespace CODE_Frontend.Modules
 {
