+++ 01_CODE_Frontend/ViewModel/ItemViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using CODE_GameLib.Interfaces.RoomObjects;
-using CODE_GameLib.Interfaces.RoomObjects.BoobyTraps;
-using CODE_GameLib.Interfaces.RoomObjects.Wearable;
+using CODE_GameLib.Objects.RoomObjects;
+using CODE_GameLib.Objects.RoomObjects.BoobyTraps;
+using CODE_GameLib.Objects.RoomObjects.Wearable;
 
 namespace CODE_Frontend.ViewModel
 {
