+++ 01_CODE_Frontend/ViewModel/PlayerViewModel.cs
@@ -1,4 +1,4 @@
-﻿using CODE_GameLib.Interfaces.Entity;
+﻿using CODE_GameLib.Objects.Entity;
 
 namespace CODE_Frontend.ViewModel
 {
