+++ 01_CODE_Frontend/ViewModel/RoomViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Linq;
-using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.Entity;
+using CODE_GameLib.Objects;
+using CODE_GameLib.Objects.Entity;
 
 namespace CODE_Frontend.ViewModel
 {
