--- 02_CODE_GameLib/Interfaces/Doors/IClosingDoor.cs
@@ -1,6 +1,0 @@
-namespace CODE_GameLib.Interfaces.Doors
-{
-    public interface IClosingDoor : IDoor
-    {
-    }
-}
