--- 02_CODE_GameLib/Interfaces/Doors/IColoredDoor.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace CODE_GameLib.Interfaces.Doors
-{
-    public interface IColoredDoor : IDoor
-    {
-        public ConsoleColor Color { get; }
-    }
-}
