--- 02_CODE_GameLib/Interfaces/Doors/IDoor.cs
@@ -1,10 +1,0 @@
-using CODE_GameLib.Interfaces.Entity;
-
-namespace CODE_GameLib.Interfaces.Doors
-{
-    public interface IDoor
-    {
-        public bool Opened { get; set; }
-        public bool PassThru(IEntity entity);
-    }
-}
