--- 02_CODE_GameLib/Interfaces/Doors/IToggleDoor.cs
@@ -1,7 +1,0 @@
-namespace CODE_GameLib.Interfaces.Doors
-{
-    public interface IToggleDoor : IDoor
-    {
-        public void ActivateToggleDoor();
-    }
-}
