--- 02_CODE_GameLib/Interfaces/IBaseObservable.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace CODE_GameLib.Interfaces
-{
-    public interface IBaseObservable<out T> : IObservable<T>
-    {
-    }
-}
