--- 02_CODE_GameLib/Interfaces/RoomObjects/BoobyTraps/IBoobyTrap.cs
@@ -1,7 +1,0 @@
-namespace CODE_GameLib.Interfaces.RoomObjects.BoobyTraps
-{
-    public interface IBoobyTrap : IRoomObject
-    {
-        public int Damage { get; }
-    }
-}
