--- 02_CODE_GameLib/Interfaces/RoomObjects/BoobyTraps/IDisappearingTrap.cs
@@ -1,6 +1,0 @@
-namespace CODE_GameLib.Interfaces.RoomObjects.BoobyTraps
-{
-    public interface IDisappearingTrap : IBoobyTrap
-    {
-    }
-}
