--- 02_CODE_GameLib/Interfaces/RoomObjects/IConveyorBelt.cs
@@ -1,9 +1,0 @@
-using CODE_GameLib.Enums;
-
-namespace CODE_GameLib.Interfaces.RoomObjects
-{
-    public interface IConveyorBelt : IRoomObject
-    {
-        Direction Direction { get; }
-    }
-}
