--- 02_CODE_GameLib/Interfaces/RoomObjects/IPortal.cs
@@ -1,11 +1,0 @@
-using CODE_GameLib.Interfaces.Entity;
-
-namespace CODE_GameLib.Interfaces.RoomObjects
-{
-    public interface IPortal : IRoomObject
-    {
-        ILocation Destination { get; }
-
-        public bool UsePortal(IEntity entity);
-    }
-}
