--- 02_CODE_GameLib/Interfaces/RoomObjects/IPressurePlate.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace CODE_GameLib.Interfaces.RoomObjects
-{
-    public interface IPressurePlate : IRoomObject
-    {
-        public void ActivatePressurePlate(IEnumerable<IConnection> connections);
-    }
-}
