--- 02_CODE_GameLib/Interfaces/RoomObjects/IRoomObject.cs
@@ -1,8 +1,0 @@
-namespace CODE_GameLib.Interfaces.RoomObjects
-{
-    public interface IRoomObject
-    {
-        public int X { get; }
-        public int Y { get; }
-    }
-}
