--- 02_CODE_GameLib/Interfaces/RoomObjects/Wearable/IKey.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace CODE_GameLib.Interfaces.RoomObjects.Wearable
-{
-    public interface IKey : IWearable
-    {
-        public ConsoleColor Color { get; }
-    }
-}
