--- 02_CODE_GameLib/Interfaces/RoomObjects/Wearable/ISankaraStone.cs
@@ -1,6 +1,0 @@
-namespace CODE_GameLib.Interfaces.RoomObjects.Wearable
-{
-    public interface ISankaraStone : IWearable
-    {
-    }
-}
