+++ 02_CODE_GameLib/Objects/Doors/ClosingDoor.cs
@@ -1,5 +1,4 @@
-using CODE_GameLib.Interfaces.Doors;
-using CODE_GameLib.Interfaces.Entity;
+using CODE_GameLib.Objects.Entity;
 
 namespace CODE_GameLib.Objects.Doors
 {
@@ -16,5 +15,9 @@ namespace CODE_GameLib.Objects.Doors
             Opened = false;
             return true;
         }
+    }
+
+    public interface IClosingDoor : IDoor
+    {
     }
 }
