+++ 02_CODE_GameLib/Objects/Doors/ColoredDoor.cs
@@ -1,8 +1,7 @@
 using System;
 using System.Linq;
-using CODE_GameLib.Interfaces.Doors;
-using CODE_GameLib.Interfaces.Entity;
-using CODE_GameLib.Interfaces.RoomObjects.Wearable;
+using CODE_GameLib.Objects.Entity;
+using CODE_GameLib.Objects.RoomObjects.Wearable;
 
 namespace CODE_GameLib.Objects.Doors
 {
@@ -23,5 +22,10 @@ namespace CODE_GameLib.Objects.Doors
                 Opened = true;
             return Opened;
         }
+    }
+
+    public interface IColoredDoor : IDoor
+    {
+        public ConsoleColor Color { get; }
     }
 }
