+++ 02_CODE_GameLib/Objects/Doors/Door.cs
@@ -1,5 +1,4 @@
-using CODE_GameLib.Interfaces.Doors;
-using CODE_GameLib.Interfaces.Entity;
+using CODE_GameLib.Objects.Entity;
 
 namespace CODE_GameLib.Objects.Doors
 {
@@ -11,5 +10,11 @@ namespace CODE_GameLib.Objects.Doors
         {
             return Opened;
         }
+    }
+
+    public interface IDoor
+    {
+        public bool Opened { get; set; }
+        public bool PassThru(IEntity entity);
     }
 }
