+++ 02_CODE_GameLib/Objects/Entity/Player.cs
@@ -1,8 +1,6 @@
 using System.Collections.Generic;
 using CODE_GameLib.Enums;
-using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.Entity;
-using CODE_GameLib.Interfaces.RoomObjects.Wearable;
+using CODE_GameLib.Objects.RoomObjects.Wearable;
 
 namespace CODE_GameLib.Objects.Entity
 {
@@ -23,5 +21,13 @@ namespace CODE_GameLib.Objects.Entity
             NotifyObservers(this);
             return true;
         }
+    }
+
+    public interface IPlayer : IEntity
+    {
+        public List<IWearable> Inventory { get; }
+        public List<Cheat> Cheats { get; }
+        public ILocation StartLocation { get; }
+        public bool AddToInventory(IWearable wearable);
     }
 }
