+++ 02_CODE_GameLib/Objects/Game.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using CODE_GameLib.Enums;
-using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.Entity;
+using CODE_GameLib.Objects.Entity;
 using CODE_GameLib.Objects.Observers;
 
 namespace CODE_GameLib.Objects
@@ -52,5 +51,16 @@ namespace CODE_GameLib.Objects
         {
             Updated?.Invoke(this, this);
         }
+    }
+
+    public interface IGame
+    {
+        public bool Quit { get; }
+        public bool Won { get; }
+        public IPlayer Player { get; }
+        public event EventHandler<Game> Updated;
+        public void Update();
+        public void Tick(TickData tickData);
+        public void Destroy(bool winGame);
     }
 }
