+++ 02_CODE_GameLib/Objects/RoomObjects/BoobyTraps/DisappearingTrap.cs
@@ -1,11 +1,13 @@
-using CODE_GameLib.Interfaces.RoomObjects.BoobyTraps;
-
-namespace CODE_GameLib.Objects.RoomObjects
+namespace CODE_GameLib.Objects.RoomObjects.BoobyTraps
 {
     public class DisappearingTrap : BoobyTrap, IDisappearingTrap
     {
         public DisappearingTrap(int x, int y, int damage) : base(x, y, damage)
         {
         }
+    }
+
+    public interface IDisappearingTrap : IBoobyTrap
+    {
     }
 }
