+++ 02_CODE_GameLib/Objects/RoomObjects/ConveyorBelt.cs
@@ -1,12 +1,19 @@
 using CODE_GameLib.Enums;
-using CODE_GameLib.Interfaces.RoomObjects;
 
 namespace CODE_GameLib.Objects.RoomObjects
 {
     public class ConveyorBelt : RoomObject, IConveyorBelt
     {
-        public ConveyorBelt(int x, int y, Direction direction) : base(x, y) => Direction = direction;
-        
+        public ConveyorBelt(int x, int y, Direction direction) : base(x, y)
+        {
+            Direction = direction;
+        }
+
         public Direction Direction { get; }
+    }
+
+    public interface IConveyorBelt : IRoomObject
+    {
+        Direction Direction { get; }
     }
 }
