+++ 02_CODE_GameLib/Objects/RoomObjects/Wearable/Key.cs
@@ -1,12 +1,19 @@
 using System;
-using CODE_GameLib.Interfaces.RoomObjects.Wearable;
 
-namespace CODE_GameLib.Objects.RoomObjects
+namespace CODE_GameLib.Objects.RoomObjects.Wearable
 {
     public class Key : RoomObject, IKey
     {
-        public Key(int x, int y, ConsoleColor color) : base(x, y) => Color = color;
-        
+        public Key(int x, int y, ConsoleColor color) : base(x, y)
+        {
+            Color = color;
+        }
+
+        public ConsoleColor Color { get; }
+    }
+
+    public interface IKey : IWearable
+    {
         public ConsoleColor Color { get; }
     }
 }
