+++ 02_CODE_GameLib/Objects/RoomObjects/Wearable/SankaraStone.cs
@@ -1,11 +1,13 @@
-using CODE_GameLib.Interfaces.RoomObjects.Wearable;
-
-namespace CODE_GameLib.Objects.RoomObjects
+namespace CODE_GameLib.Objects.RoomObjects.Wearable
 {
     public class SankaraStone : RoomObject, ISankaraStone
     {
         public SankaraStone(int x, int y) : base(x, y)
         {
         }
+    }
+
+    public interface ISankaraStone : IWearable
+    {
     }
 }
