+++ 03_CODE_PersistenceLib/Factories/EntityLocationFactory.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using CODE_GameLib.Interfaces;
 using CODE_GameLib.Objects;
 using Newtonsoft.Json.Linq;
