+++ 03_CODE_PersistenceLib/Factories/PlayerFactory.cs
@@ -1,5 +1,4 @@
-﻿using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.Entity;
+﻿using CODE_GameLib.Objects;
 using CODE_GameLib.Objects.Entity;
 using Newtonsoft.Json.Linq;
