+++ 03_CODE_PersistenceLib/Factories/PortalFactory.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
 using System.Linq;
-using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.RoomObjects;
 using CODE_GameLib.Objects;
 using CODE_GameLib.Objects.RoomObjects;
 using Newtonsoft.Json.Linq;
