+++ 03_CODE_PersistenceLib/Factories/RoomFactory.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using CODE_GameLib.Interfaces;
-using CODE_GameLib.Interfaces.RoomObjects;
 using CODE_GameLib.Objects;
+using CODE_GameLib.Objects.RoomObjects;
 using Newtonsoft.Json.Linq;
 
 namespace CODE_PersistenceLib.Factories
